Extract Notes component
