+++ client/src/App.css
@@ -309,49 +309,86 @@ button.password--submit {
   position: relative;
 }
 
-.list-wrapper svg {
-  position: absolute;
-  font-size: 2.5rem;
-  transition: transform 200ms ease;
-}
-
 .list-wrapper span {
   overflow-wrap: break-word;
   word-wrap: break-word;
   word-break: break-word;
 }
 
-.notes-container {
+.list-wrapper svg {
   position: absolute;
   z-index: 1;
+  right: 5%;
+  top: 4px;
+  font-size: 1.875rem;
+  transition: transform 200ms ease;
+  opacity: 0.7;
+}
+
+.notes-container {
   background: var(--paper-color) var(--paper-texture);
+  padding: 1rem;
+  box-shadow: 1px 1px 1px 1px #381a1480, 2px 1px 0 0 var(--paper-color);
+}
+
+.notes-border {
+  position: absolute;
+  width: 95%;
+  z-index: 3;
+  background: var(--leather-color) var(--leather-texture);
+  padding: 0.4rem;
+  border-radius: 0.5rem;
+  box-shadow: 4px 4px 8px rgba(0, 0, 0, 0.4);
+}
+
+.notes-border ~ .menu-backdrop {
+  z-index: 2;
 }
 
 .revealnotes-enter,
 .revealnotes-exit-done {
   transform: scaleY(0);
+}
+
+.revealnotes-enter > .notes-container,
+.revealnotes-exit-done > .notes-container {
   opacity: 0;
 }
 
 .revealnotes-enter-active,
 .revealnotes-exit-active {
   transform-origin: top;
-  transition: transform 300ms ease, opacity 200ms ease;
+  transition: transform 300ms ease;
+}
+
+.revealnotes-enter-active > .notes-container,
+.revealnotes-exit-active > .notes-container {
+  transition: opacity 200ms ease;
 }
 
 .revealnotes-enter-active {
   transform: scaleY(1);
+}
+
+.revealnotes-enter-active > .notes-container {
   opacity: 1;
 }
 
 .revealnotes-exit-active {
   transform: scaleY(0);
+}
+
+.revealnotes-exit-active > .notes-container {
   opacity: 0;
 }
 
 /* .revealnotes-exit, */
 .revealnotes-enter-done {
   transform: scaleY(1);
+  opacity: 1;
+}
+
+.revealnotes-enter-done > .notes-container {
   opacity: 1;
 }
